.bash_profile updates and refactoring
Preparing to implement more modular system for configuration and inheritance. For this reason, I'm rearranging the .bash_profile in an effort to allow for variables to be easily overridden by platform-specific bash_profile includes.

Also includes updated logic for bash prompt "git status" call which will work on more versions of git (uses the shortened output and checks if it is null, rather than for a specific last line of output).